fix: resolve Ktor compile errors (lowercase post, long setex, environment.monitor)

+++ src/main/kotlin/com/example/challenge/adapters/egress/cache/RedisCacheRepository.kt
@@ -39,7 +39,7 @@ class RedisCacheRepository : CachePort {
         try {
             val json = Json.encodeToString(rate)
             RedisConfig.getResource().use { jedis ->
-                jedis.setex(key, cacheDurationSeconds, json)
+                jedis.setex(key, cacheDurationSeconds.toLong(), json)
                 logger.info("Saved tax rate to Redis cache with key: {}", key)
             }
         } catch (e: Exception) {
